CHORD-node: Extract TestChordIdentifierScope
