+++ hephaistus/main-theme/distribution-system/chord/playground/nullist0/lib/src/test/kotlin/org/lia/dht/nullist0/chord/model/ChordIdentifierScopeTest.kt
@@ -1,6 +1,7 @@
 package org.lia.dht.nullist0.chord.model
 
 import org.junit.jupiter.api.Test
+import org.lia.dht.nullist0.chord.fake.TestChordIdentifierScope
 import kotlin.test.assertEquals
 
 class ChordIdentifierScopeTest {
@@ -44,13 +45,5 @@ class ChordIdentifierScopeTest {
 
             assertEquals(expected, value, "$v is failed with $pair: expected=$expected")
         }
-    }
-
-    class TestChordIdentifierScope: ChordIdentifierScope<Int>() {
-        override fun ChordNode<Int>.toId(): Int  = error("Not supported")
-        override fun Int.toId(): Int  = error("Not supported")
-        override fun bitSize(): Int  = error("Not supported")
-        override fun Int.compareTo(other: Int): Int  = compareTo(other)
-        override fun Int.plus(other: Int): Int = error("Not supported")
     }
 }
